% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumadosnumeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma dos números}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{Primer valor numérico de la suma}

\item{y}{Segundo valor numérico de la suma}
}
\value{
Devuelve la suma de los dos valores
}
\description{
Esta función acepta dos números y los suma.
}
\note{
Esta función acepta como parámetros sólamente valores únicos.
En el caso de que se proporcionen más o menos parametros de los dos aceptados,
la función dará error y parará.
}
\examples{
> suma.dos.numeros(2,2)
> suma.dos.numeros(2.5)              # Devuelve error
> suma.dos.numeros(c(2,2), c(2,5))   # Devuelve error
}

